requirements.txt
aiofiles==23.2.1
fastapi==0.99.1
sqlmodel==0.0.24
passlib==1.7.4
python-jose==3.4.0
Jinja2==3.1.2
mangum==0.17.0
openai==1.3.6
python-multipart==0.0.6